add fmpz_vec
